backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
pandas==2.3.3
openpyxl==3.1.5
python-dotenv==1.1.1
google-generativeai==0.8.5
requests==2.32.5
Werkzeug==3.1.4
gunicorn==23.0.0
python-multipart==0.0.20